+++ Sources/Sentry/include/SentryDateUtil.h
@@ -13,6 +13,8 @@ SENTRY_NO_INIT
 
 + (NSDate *_Nullable)getMaximumDate:(NSDate *_Nullable)first andOther:(NSDate *_Nullable)second;
 
++ (long)millisecondsSince1970:(NSDate *)date;
+
 @end
 
 NS_ASSUME_NONNULL_END