+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## Unreleased
 
+### Features
+
+- Add Session Replay, which is **still experimental**. (#3625)
+
 ### Fixes
 
 - Crash due to a background call to -[UIApplication applicationState] (#3855)
+- Save framework without UIKit/AppKit as Github Asset for releases (#3858) 
+- Crash due to a background call to -[UIApplication applicationState] (#3836)
 
 ## 8.24.0